unit tests completed with mockMvc

+++ books/src/main/java/com/books/books/controller/BookController.java
@@ -1,79 +1,67 @@
 package com.books.books.controller;
 
 import com.books.books.controller.dto.BookDto;
-import com.books.books.entity.Book;
 import com.books.books.entity.Genre;
 import com.books.books.sevice.BookService;
-import com.books.books.sevice.BookServiceImpl;
-import jakarta.persistence.GeneratedValue;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import jakarta.validation.Valid;
 
-import java.util.ArrayList;
+
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/book")
 public class BookController {
 
-    BookServiceImpl bookService;
+    BookService bookService;
 
-    public BookController(BookServiceImpl bookService) {
+    public BookController(BookService bookService) {
         this.bookService = bookService;
     }
 
     @GetMapping("allbooks")
     public ResponseEntity<?> listAllBooks() {
-        List<Book> bookList = bookService.listAllBooks();
-        List<BookDto> bookDtoList = new ArrayList<>();
-        BookDto bookDto;
-        for (Book b : bookList) {
-            bookDto = BookDto.builder()
-                    .id(b.getId())
-                    .author(b.getAuthor())
-                    .genre(b.getGenre())
-                    .name(b.getName())
-                    .price(b.getPrice())
-                    .build();
-            bookDtoList.add(bookDto);
-        }
-        return ResponseEntity.ok(bookDtoList);
+        List<BookDto> bookListDto = bookService.listAllBooks();
+        return ResponseEntity.ok(bookListDto);
     }
 
     @GetMapping("listbooksbygender/{value}")
     public ResponseEntity<?> listBooksByGenre(@PathVariable String value) {
         Genre genre = Genre.convertValue(value);
         if (genre != null) {
-            List<Book> bookList = bookService.listAllBooksByGenre(genre);
-            List<BookDto> bookDtoList = new ArrayList<>();
-            BookDto bookDto;
-            for (Book b : bookList) {
-                bookDto = BookDto.builder()
-                        .id(b.getId())
-                        .author(b.getAuthor())
-                        .genre(b.getGenre())
-                        .name(b.getName())
-                        .price(b.getPrice())
-                        .build();
-                bookDtoList.add(bookDto);
-            }
-            return ResponseEntity.ok(bookDtoList);
+            List<BookDto> bookListDto = bookService.listAllBooksByGenre(genre);
+            return ResponseEntity.ok(bookListDto);
         } else return ResponseEntity.badRequest().body("Genre not exist");
+    }
+
+    @PutMapping("/updatebook/{id}")
+    public ResponseEntity<?> updateBook (@RequestBody BookDto bookDto,@PathVariable Long id){
+        Optional<BookDto> bookDtoActualized = bookService.updateBook(bookDto,id);
+
+        if (bookDtoActualized.isPresent()){
+            return ResponseEntity.ok(bookDtoActualized.get());
+        }
+        return ResponseEntity.badRequest().body("incorrect Data");
     }
 
     @PostMapping("/newbook")
     public ResponseEntity<?> newBook(@Valid @RequestBody BookDto bookDto) {
-       // if (bookService.dtoIsOk(bookDto)) {
+        try {
             return ResponseEntity.ok(bookService.saveBook(bookDto));
-        //}
-        //return ResponseEntity.badRequest().body("Incorrect data");
+
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body("incomplete fields");
+        }
     }
 
     @DeleteMapping("deletebook/{id}")
     public ResponseEntity<?> deleteBook(@PathVariable Long id) {
-        bookService.deleteBookById(id);
-        return ResponseEntity.ok("Registro Eliminado");
+        if(bookService.idExist(id)) {
+            bookService.deleteBookById(id);
+            return ResponseEntity.ok("Registro Eliminado");
+        }
+        else return ResponseEntity.badRequest().body("Id not exist");
     }
 }
